Finished addressing comments.
