+++ src/main/java/org/broadinstitute/hellbender/tools/funcotator/DataSourceFuncotationFactory.java
@@ -2,7 +2,6 @@ package org.broadinstitute.hellbender.tools.funcotator;
 
 import htsjdk.tribble.Feature;
 import htsjdk.variant.variantcontext.VariantContext;
-import org.broadinstitute.hellbender.engine.FeatureContext;
 import org.broadinstitute.hellbender.engine.ReferenceContext;
 
 import java.util.List;
@@ -11,17 +10,17 @@ import java.util.List;
  * An abstract class to allow for the creation of a {@link Funcotation} for a given data source.
  * Created by jonn on 8/30/17.
  */
-public abstract class DataSourceFuncotationFactory implements AutoCloseable {
+public interface DataSourceFuncotationFactory extends AutoCloseable {
 
     /**
      * Perform cleanup tasks for this {@link DataSourceFuncotationFactory}.
      */
-    public void close() {}
+    default void close() {}
 
     /**
      * @return An ordered list of the names of annotations that this Data Source supports.
      */
-    public abstract List<String> getSupportedFuncotationFields();
+    List<String> getSupportedFuncotationFields();
 
     /**
      * Creates a {@link List} of {@link Funcotation} for the given {@code variant}, {@code referenceContext}, and {@code featureContext}.
@@ -30,5 +29,5 @@ public abstract class DataSourceFuncotationFactory implements AutoCloseable {
      * @param featureList {@link List} of {@link Feature} corresponding to the given {@code variant}.
      * @return {@link List} of {@link Funcotation} given the {@code variant}, {@code referenceContext}, and {@code featureContext}.  This should never be empty.
      */
-    public abstract List<Funcotation> createFuncotations(final VariantContext variant, final ReferenceContext referenceContext, final List<Feature> featureList);
+    List<Funcotation> createFuncotations(final VariantContext variant, final ReferenceContext referenceContext, final List<Feature> featureList);
 }
