+++ src/main/java/org/broadinstitute/hellbender/tools/funcotator/FuncotatorUtils.java
@@ -1,54 +1,4 @@
-/*
-* Copyright 2012-2016 Broad Institute, Inc.
-* 
-* Permission is hereby granted, free of charge, to any person
-* obtaining a copy of this software and associated documentation
-* files (the "Software"), to deal in the Software without
-* restriction, including without limitation the rights to use,
-* copy, modify, merge, publish, distribute, sublicense, and/or sell
-* copies of the Software, and to permit persons to whom the
-* Software is furnished to do so, subject to the following
-* conditions:
-* 
-* The above copyright notice and this permission notice shall be
-* included in all copies or substantial portions of the Software.
-* 
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-* OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-* NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-* HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-* WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-* FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR
-* THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-*/
-
 package org.broadinstitute.hellbender.tools.funcotator;
-
-/*
- * Copyright (c) 2010 The Broad Institute
- *
- * Permission is hereby granted, free of charge, to any person
- * obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without
- * restriction, including without limitation the rights to use,
- * copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following
- * conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- * OTHER DEALINGS IN THE SOFTWARE.
- */
 
 import htsjdk.samtools.reference.ReferenceSequence;
 import htsjdk.samtools.util.Locatable;
@@ -59,17 +9,15 @@ import org.apache.log4j.Logger;
 import org.broadinstitute.hellbender.engine.ReferenceContext;
 import org.broadinstitute.hellbender.exceptions.GATKException;
 import org.broadinstitute.hellbender.exceptions.UserException;
-import org.broadinstitute.hellbender.utils.GATKProtectedVariantContextUtils;
 import org.broadinstitute.hellbender.utils.SimpleInterval;
 import org.broadinstitute.hellbender.utils.Utils;
 import org.broadinstitute.hellbender.utils.param.ParamUtils;
 import org.broadinstitute.hellbender.utils.read.ReadUtils;
+import org.broadinstitute.hellbender.utils.variant.GATKVariantContextUtils;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
+import java.util.*;
 
-public class FuncotatorUtils {
+public final class FuncotatorUtils {
 
     private final static Logger logger = Logger.getLogger(FuncotatorUtils.class);
 
@@ -79,21 +27,30 @@ public class FuncotatorUtils {
      */
     private FuncotatorUtils() {}
 
-    private static final HashMap<String, AminoAcid> tableByCodon = new HashMap<>(AminoAcid.values().length);
-    private static final HashMap<String, AminoAcid> tableByCode = new HashMap<>(AminoAcid.values().length);
-    private static final HashMap<String, AminoAcid> tableByLetter = new HashMap<>(AminoAcid.values().length);
+    private static final Map<String, AminoAcid> tableByCodon;
+    private static final Map<String, AminoAcid> tableByCode;
+    private static final Map<String, AminoAcid> tableByLetter;
 
     /**
      * Initialize our hashmaps of lookup tables:
      */
     static {
+
+        final HashMap<String, AminoAcid> mapByCodon  = new HashMap<>(AminoAcid.values().length);
+        final HashMap<String, AminoAcid> mapByLetter = new HashMap<>(AminoAcid.values().length);
+        final HashMap<String, AminoAcid> mapByCode   = new HashMap<>(AminoAcid.values().length);
+
         for ( final AminoAcid acid : AminoAcid.values() ) {
-            tableByCode.put(acid.getCode(),acid);
-            tableByLetter.put(acid.getLetter(), acid);
+            mapByCode.put(acid.getCode(),acid);
+            mapByLetter.put(acid.getLetter(), acid);
             for ( final String codon : acid.getCodons() ) {
-                tableByCodon.put(codon,acid);
+                mapByCodon.put(codon,acid);
             }
         }
+
+        tableByCodon = Collections.unmodifiableMap(mapByCodon);
+        tableByCode = Collections.unmodifiableMap(mapByCode);
+        tableByLetter = Collections.unmodifiableMap(mapByLetter);
     }
 
     /**
@@ -414,17 +371,17 @@ public class FuncotatorUtils {
         final StringBuilder alt = new StringBuilder();
 
         // Handle the ONP case:
-        if ( GATKProtectedVariantContextUtils.isOnp(alignedRefAllele, alignedAltAllele) ) {
+        if ( GATKVariantContextUtils.isOnp(alignedRefAllele, alignedAltAllele) ) {
             return getCodonChangeStringForOnp( alignedRefAllele, alignedAltAllele,
                     seqComp.getAlignedCodingSequenceAlleleStart(), seqComp.getAlignedReferenceAlleleStop() );
         }
         // Handle the insertion case:
-        else if ( GATKProtectedVariantContextUtils.isInsertion(alignedRefAllele, alignedAltAllele) ) {
+        else if ( GATKVariantContextUtils.isInsertion(alignedRefAllele, alignedAltAllele) ) {
 
             final String nextRefCodon = getNextReferenceCodon(seqComp.getReferenceCodingSequence(), seqComp.getAlignedCodingSequenceAlleleStart(), seqComp.getAlignedReferenceAlleleStop(), seqComp.getStrand());
 
             // Check for frame shift syntax first:
-            if ( GATKProtectedVariantContextUtils.isFrameshift(alignedRefAllele, alignedAltAllele) ) {
+            if ( GATKVariantContextUtils.isFrameshift(alignedRefAllele, alignedAltAllele) ) {
                 return getCodonChangeStringForInsertionFrameshift(
                         seqComp.getCodingSequenceAlleleStart(), seqComp.getAlignedCodingSequenceAlleleStart(), seqComp.getAlignedReferenceAlleleStop(),
                         nextRefCodon, seqComp.getAlignedReferenceAllele() );
@@ -658,7 +615,7 @@ public class FuncotatorUtils {
         // (i.e. a base is added to the beginning of the string), we can ignore the first codon (3 bases) of the
         // reference:
         String decorator = "del";
-        if ( GATKProtectedVariantContextUtils.isFrameshift(alignedRefAllele, alignedAltAllele) ) {
+        if ( GATKVariantContextUtils.isFrameshift(alignedRefAllele, alignedAltAllele) ) {
             decorator = "fs";
         }
 
@@ -811,11 +768,11 @@ public class FuncotatorUtils {
         final Integer protChangeEndPos = seqComp.getProteinChangeEndPosition();
 
         // Check for the ONP case:
-        if ( GATKProtectedVariantContextUtils.isOnp(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
+        if ( GATKVariantContextUtils.isOnp(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
             return getProteinChangeStringForOnp(refAaSeq, altAaSeq, protChangeStartPos, protChangeEndPos);
         }
         // Check for the Insertion case:
-        else if ( GATKProtectedVariantContextUtils.isInsertion(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
+        else if ( GATKVariantContextUtils.isInsertion(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
 
             Utils.nonNull(seqComp.getCodingSequenceAlleleStart());
             Utils.nonNull(seqComp.getReferenceCodingSequence());
@@ -827,7 +784,7 @@ public class FuncotatorUtils {
             // We must also know the Amino Acid sequence for it:
             final String nextRefAaSeq = createAminoAcidSequence(nextRefCodon);
 
-            if ( GATKProtectedVariantContextUtils.isFrameshift(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
+            if ( GATKVariantContextUtils.isFrameshift(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
                 return getProteinChangeStringForInsertionFrameshift(
                         seqComp.getCodingSequenceAlleleStart(),
                         refAaSeq,
@@ -936,7 +893,7 @@ public class FuncotatorUtils {
         // (i.e. a base is added to the beginning of the string), we can ignore the first amino acid (3 bases) of the
         // reference:
         String decorator = "del";
-        if ( GATKProtectedVariantContextUtils.isFrameshift(referenceAllele, alternateAllele) ) {
+        if ( GATKVariantContextUtils.isFrameshift(referenceAllele, alternateAllele) ) {
             decorator = "fs";
         }
         return "p." + referenceAminoAcidSequence.substring(1) + (protChangeStartPos + 1) + decorator;
@@ -1049,7 +1006,7 @@ public class FuncotatorUtils {
         Utils.nonNull(seqComp.getAlternateAllele());
 
         // Check for ONP:
-        if ( GATKProtectedVariantContextUtils.isOnp(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
+        if ( GATKVariantContextUtils.isOnp(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
             if (seqComp.getAlternateAllele().length() > 1) {
                 return "c." + seqComp.getCodingSequenceAlleleStart() + "_" + (seqComp.getCodingSequenceAlleleStart() + seqComp.getReferenceAllele().length() - 1) +
                         seqComp.getReferenceAllele() + ">" + seqComp.getAlternateAllele();
@@ -1059,7 +1016,7 @@ public class FuncotatorUtils {
             }
         }
         // Check for Insertion:
-        else if ( GATKProtectedVariantContextUtils.isInsertion(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
+        else if ( GATKVariantContextUtils.isInsertion(seqComp.getReferenceAllele(), seqComp.getAlternateAllele()) ) {
             return "c." + seqComp.getCodingSequenceAlleleStart() + "_" + (seqComp.getCodingSequenceAlleleStart() + 1) +
                     "ins" + seqComp.getAlternateAllele().substring(seqComp.getReferenceAllele().length()).toUpperCase();
         }
@@ -1551,364 +1508,5 @@ public class FuncotatorUtils {
             throw new GATKException("Unable to handle NONE strand.");
         }
 
-    }
-
-    /**
-     * A simple data object to hold a comparison between a reference sequence and an alternate allele.
-     */
-    public static class SequenceComparison {
-
-        /**
-         * Bases covering the region around the variant.
-         * This includes bases that would overlap any of the variant or reference bases.
-         * (i.e. if the reference allele is 'A' and the variant is 'GCGCG', this would be
-         * 'ANNNNN' where 'N' is the correct base at that position in the reference sequence).
-         * Contains {@link FuncotatorUtils.SequenceComparison#referenceWindow} bases of padding before and
-         * after the total length.
-         */
-        private String referenceBases                     = null;
-
-        /**
-         * The number of bases in {@link FuncotatorUtils.SequenceComparison#referenceBases} before the
-         * start of the reference Allele / variant.
-         */
-        private Integer referenceWindow                   = null;
-
-        /**
-         * The reference coding sequence for a the transcript of this sequence comparison.
-         * This does NOT include introns.
-         * Stored in the forward reading direction.  For NEGATIVE strand reads, must
-         * reverse complement any bases retrieved.
-         */
-        private ReferenceSequence referenceCodingSequence = null;
-
-        /**
-         * The contig on which this sequence comparison occurs.
-         */
-        private String  contig                               = null;
-
-        /**
-         * The strand on which this sequence comparison occurs.
-         */
-        private Strand  strand                               = null;
-
-        /**
-         * The position (1-based, inclusive in genome coordinates - relative to the start of
-         * {@link FuncotatorUtils.SequenceComparison#contig}) of the start of the reference allele / variant.
-         */
-        private Integer alleleStart                          = null;
-
-        /**
-         * The position (1-based, inclusive) in transcript coordinates relative to the start of
-         * the transcript of start of the reference allele / variant.
-         */
-        private Integer transcriptAlleleStart                = null;
-
-        /**
-         * The position (1-based, inclusive) in coding sequence coordinates relative to the start of
-         * the coding region of the transcript of the start of the reference allele / variant.
-         * This location is obtained by concatenating the exons together and counting from the start of
-         * that sequence to where the reference allele / variant begins.
-         */
-        private Integer codingSequenceAlleleStart            = null;
-
-        /**
-         * The in-frame position (1-based, inclusive) in coding sequence coordinates relative to the start of
-         * the coding region of the transcript of the start of the first codon containing the reference allele / variant.
-         * This location is obtained by concatenating the exons together and counting from the start of
-         * that sequence to where the reference allele / variant begins, then moving backwards to an in-frame
-         * position, if necessary.
-         */
-        private Integer alignedCodingSequenceAlleleStart     = null;
-
-        /**
-         * The position (1-based, inclusive in genome coordinates - relative to the start of
-         * {@link FuncotatorUtils.SequenceComparison#contig}) of the start of the exon that contains the
-         * variant in this {@link FuncotatorUtils.SequenceComparison}.
-         */
-        private Integer exonStartPosition                    = null;
-
-        /**
-         * The position (1-based, inclusive in genome coordinates - relative to the start of
-         * {@link FuncotatorUtils.SequenceComparison#contig}) of the end of the exon that contains the
-         * variant in this {@link FuncotatorUtils.SequenceComparison}.
-         */
-        private Integer exonEndPosition                      = null;
-
-        /**
-         * The start position (1-based, inclusive) of the Protein Change for the alleles in this {@link SequenceComparison}.
-         * This is computed using {@link SequenceComparison#alignedCodingSequenceAlleleStart}.
-         * See {@link FuncotatorUtils#getProteinChangePosition} for more details.
-         */
-        private Integer proteinChangeStartPosition           = null;
-
-        /**
-         * The end position (1-based, inclusive) of the Protein Change for the alleles in this {@link SequenceComparison}.
-         * This is computed by using {@link SequenceComparison#alignedCodingSequenceAlleleStart} and the length of
-         * {@link SequenceComparison#alignedCodingSequenceAlternateAllele}.
-         * See {@link FuncotatorUtils#getProteinChangeEndPosition} for more details.
-         */
-        private Integer proteinChangeEndPosition             = null;
-
-        /**
-         * A string representation of the reference allele.
-         */
-        private String  referenceAllele                      = null;
-        /**
-         * An in-frame sequence of bases that overlaps the given reference allele based on the raw reference genome.
-         * (i.e. This includes INTRONS.)
-         */
-        private String  alignedReferenceAllele               = null;
-        /**
-         * An in-frame sequence of bases that overlaps the given reference allele for the coding region only.
-         * (i.e. This includes ONLY EXONS.)
-         */
-        private String  alignedCodingSequenceReferenceAllele = null;
-
-        /**
-         * The in-frame position (1-based, inclusive) of the last base of the last codon that contains the reference
-         * allele relative to the start of the coding sequence.
-         * All codons containing the reference allele can be extracted from the coding sequence using
-         * {@link SequenceComparison#alignedCodingSequenceAlleleStart} and {@link SequenceComparison#alignedReferenceAlleleStop}.
-         */
-        private Integer alignedReferenceAlleleStop           = null;
-
-        /**
-         * The amino acid sequence as coded by the in-frame reference coding sequence ({@link SequenceComparison#alignedCodingSequenceReferenceAllele}.
-         */
-        private String  referenceAminoAcidSequence           = null;
-
-        /**
-         * A string representation of the alternate allele.
-         */
-        private String  alternateAllele                      = null;
-
-        /**
-         * An in-frame sequence of bases that includes the entire alternate allele based on the reference genome.
-         * May span multiple codons.
-         * May include intron bases.
-         */
-        private String  alignedAlternateAllele               = null;
-
-        /**
-         * An in-frame sequence of bases that includes the entire alternate allele based on the coding sequence reference.
-         * May span multiple codons.
-         */
-        private String  alignedCodingSequenceAlternateAllele = null;
-
-        /**
-         * The in-frame position (1-based, inclusive) of the last base of the last codon that contains the alternate
-         * allele relative to the start of the coding sequence.
-         * All codons containing the alternate allele can be extracted from the coding sequence using
-         * {@link SequenceComparison#alignedCodingSequenceAlleleStart} and {@link SequenceComparison#alignedAlternateAlleleStop}.
-         */
-        private Integer alignedAlternateAlleleStop           = null;
-
-        /**
-         * The amino acid sequence as coded by the in-frame alternate coding sequence ({@link SequenceComparison#alignedCodingSequenceAlternateAllele}.
-         */
-        private String  alternateAminoAcidSequence           = null;
-
-        // =============================================================================================================
-
-        public String getReferenceBases() {
-            return referenceBases;
-        }
-
-        public void setReferenceBases(final String referenceBases) {
-            this.referenceBases = referenceBases;
-        }
-
-        public Integer getReferenceWindow() {
-            return referenceWindow;
-        }
-
-        public void setReferenceWindow(final Integer referenceWindow) {
-            this.referenceWindow = referenceWindow;
-        }
-
-        /**
-         * Return the {@link ReferenceSequence} containing the coding region for the transcript of this {@link SequenceComparison}.
-         * This does NOT include introns.
-         * The reference sequence is stored in the forward reading direction.
-         * For NEGATIVE strand reads, must reverse complement any bases retrieved.
-         */
-        public ReferenceSequence getReferenceCodingSequence() {
-            return referenceCodingSequence;
-        }
-
-        public void setReferenceCodingSequence(final ReferenceSequence referenceCodingSequence) {
-            this.referenceCodingSequence = referenceCodingSequence;
-        }
-
-        public String getContig() {
-            return contig;
-        }
-
-        public void setContig(final String contig) {
-            this.contig = contig;
-        }
-
-        public Strand getStrand() {
-            return strand;
-        }
-
-        public void setStrand(final Strand strand) {
-
-            if (strand == Strand.NONE) {
-                throw new GATKException("Cannot handle NONE strand.");
-            }
-
-            this.strand = strand;
-        }
-
-        public Integer getAlleleStart() {
-            return alleleStart;
-        }
-
-        public void setAlleleStart(final Integer alleleStart) {
-            this.alleleStart = alleleStart;
-        }
-
-        public Integer getTranscriptAlleleStart() {
-            return transcriptAlleleStart;
-        }
-
-        public void setTranscriptAlleleStart(final Integer transcriptAlleleStart) {
-            this.transcriptAlleleStart = transcriptAlleleStart;
-        }
-
-        public Integer getCodingSequenceAlleleStart() {
-            return codingSequenceAlleleStart;
-        }
-
-        public void setCodingSequenceAlleleStart(final Integer codingSequenceAlleleStart) {
-            this.codingSequenceAlleleStart = codingSequenceAlleleStart;
-        }
-
-        public Integer getAlignedCodingSequenceAlleleStart() {
-            return alignedCodingSequenceAlleleStart;
-        }
-
-        public void setAlignedCodingSequenceAlleleStart(final Integer alignedCodingSequenceAlleleStart) {
-            this.alignedCodingSequenceAlleleStart = alignedCodingSequenceAlleleStart;
-        }
-
-        public Integer getExonStartPosition() {
-            return exonStartPosition;
-        }
-
-        public void setExonStartPosition(final Integer exonStartPosition) {
-            this.exonStartPosition = exonStartPosition;
-        }
-
-        public Integer getExonEndPosition() {
-            return exonEndPosition;
-        }
-
-        public void setExonEndPosition(final Integer exonEndPosition) {
-            this.exonEndPosition = exonEndPosition;
-        }
-
-        public void setExonPosition( final SimpleInterval exonPosition ) {
-            this.exonStartPosition = exonPosition.getStart();
-            this.exonEndPosition = exonPosition.getEnd();
-        }
-
-        public Integer getProteinChangeStartPosition() {
-            return proteinChangeStartPosition;
-        }
-
-        public void setProteinChangeStartPosition(final Integer proteinChangeStartPosition) {
-            this.proteinChangeStartPosition = proteinChangeStartPosition;
-        }
-
-        public Integer getProteinChangeEndPosition() {
-            return proteinChangeEndPosition;
-        }
-
-        public void setProteinChangeEndPosition(final Integer proteinChangeEndPosition) {
-            this.proteinChangeEndPosition = proteinChangeEndPosition;
-        }
-
-        public String getReferenceAllele() {
-            return referenceAllele;
-        }
-
-        public void setReferenceAllele(final String referenceAllele) {
-            this.referenceAllele = referenceAllele;
-        }
-
-        public String getAlignedReferenceAllele() {
-            return alignedReferenceAllele;
-        }
-
-        public void setAlignedReferenceAllele(final String alignedReferenceAllele) {
-            this.alignedReferenceAllele = alignedReferenceAllele;
-        }
-
-        public String getAlignedCodingSequenceReferenceAllele() {
-            return alignedCodingSequenceReferenceAllele;
-        }
-
-        public void setAlignedCodingSequenceReferenceAllele(final String alignedCodingSequenceReferenceAllele) {
-            this.alignedCodingSequenceReferenceAllele = alignedCodingSequenceReferenceAllele;
-        }
-
-        public Integer getAlignedReferenceAlleleStop() {
-            return alignedReferenceAlleleStop;
-        }
-
-        public void setAlignedReferenceAlleleStop(final Integer alignedReferenceAlleleStop) {
-            this.alignedReferenceAlleleStop = alignedReferenceAlleleStop;
-        }
-
-        public String getReferenceAminoAcidSequence() {
-            return referenceAminoAcidSequence;
-        }
-
-        public void setReferenceAminoAcidSequence(final String referenceAminoAcidSequence) {
-            this.referenceAminoAcidSequence = referenceAminoAcidSequence;
-        }
-
-        public String getAlternateAllele() {
-            return alternateAllele;
-        }
-
-        public void setAlternateAllele(final String alternateAllele) {
-            this.alternateAllele = alternateAllele;
-        }
-
-        public String getAlignedAlternateAllele() {
-            return alignedAlternateAllele;
-        }
-
-        public void setAlignedAlternateAllele(final String alignedAlternateAllele) {
-            this.alignedAlternateAllele = alignedAlternateAllele;
-        }
-
-        public String getAlignedCodingSequenceAlternateAllele() {
-            return alignedCodingSequenceAlternateAllele;
-        }
-
-        public void setAlignedCodingSequenceAlternateAllele(final String alignedCodingSequenceAlternateAllele) {
-            this.alignedCodingSequenceAlternateAllele = alignedCodingSequenceAlternateAllele;
-        }
-
-        public Integer getAlignedAlternateAlleleStop() {
-            return alignedAlternateAlleleStop;
-        }
-
-        public void setAlignedAlternateAlleleStop(final Integer alignedAlternateAlleleStop) {
-            this.alignedAlternateAlleleStop = alignedAlternateAlleleStop;
-        }
-
-        public String getAlternateAminoAcidSequence() {
-            return alternateAminoAcidSequence;
-        }
-
-        public void setAlternateAminoAcidSequence(final String alternateAminoAcidSequence) {
-            this.alternateAminoAcidSequence = alternateAminoAcidSequence;
-        }
     }
 }
